Refactor into a bookdown and remove contributing
The DBPB is currently a website that links to three bookdowns.
This refactor melts all components of the DPBP site into a single
bookdown and removes the contributing page. This format may
be easier to maintain.

Delete the original DPBP completely

The plan is to start with a fresh bookdown document. Instead of adding
to the existing "collection" of bookdowns, what used to be separate books
can now just be chapters in the book.

Populate with DPBP content

As part of renovating the EDI Data Package Best Practices page, this
commit creates and populates the new bookdown skeleton with the verbatim
content from the existing site.

Transfer content from old to new DPBP

Fix the EDI logo image

Update the README

The README is unchanged from the previous DPBP. All that is different is a note at the
top stating that this is a refactored version of the original.

Make final adjustments